+++ udata/__init__.py
@@ -4,5 +4,5 @@
 uData
 '''
 
-__version__ = '2.5.2.dev'
+__version__ = '2.6.0'
 __description__ = 'Open data portal'