+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Current (in progress)
+## 2.6.0 (2021-01-25)
 
 - Add resource's description and title size limit [#2586](https://github.com/opendatateam/udata/pull/2586)
 - Add RDF catalog view for organizations [#2583](https://github.com/opendatateam/udata/pull/2583)